refactor(chat): collapse duplicate normalisers, break the emote sheet cycle, rehome stray utils
Four copies of the same username normaliser, one already divergent: modCommands
stripped the leading @ before trimming, so " @user" normalised differently
there than everywhere else. All four now call normaliseChatUsername.

EmoteSheet.tsx re-exported EmotePickerItem and emoteMenuData imported it back
through that barrel, which was the only import cycle in src. Consumers point at
emoteSheetTypes; madge reports none.

Folder layout: formatModerationSystemMessage and hydrateVisibleSevenTvAssets
each had a file sitting beside a folder of the same name, roomStateTracker sat
outside roomState/, a fixture sat outside __tests__/, and twelve utils and
hooks sat at component-folder roots.
