+++ src/components/Chat/components/ActionSheet/ActionSheet.tsx
@@ -13,7 +13,7 @@ import {
   type BottomSheetHandle,
 } from '@app/components/BottomSheet/BottomSheet';
 import { Button } from '@app/components/Button/Button';
-import { useIosActionSheet } from '@app/components/Chat/components/useIosActionSheet';
+import { useIosActionSheet } from '@app/components/Chat/hooks/useIosActionSheet';
 import { SymbolView } from '@app/components/ui/Icon/Icon';
 import { Text } from '@app/components/ui/Text/Text';
 import { theme } from '@app/styles/themes';
