+++ src/components/Chat/components/ChatComposer/components/EmoteSuggestions.tsx
@@ -16,7 +16,7 @@ import type { SanitisedEmote } from '@app/types/emote';
 import {
   suggestionRailColors,
   suggestionRailStyles,
-} from '../suggestionRailStyles';
+} from '../util/suggestionRailStyles';
 
 const EMOTE_SUGGESTION_ITEM_SIZE = 48;
