+++ src/components/Chat/components/ChatComposer/util/getCurrentWordAndType.ts
@@ -1,4 +1,4 @@
-import type { SuggestionType } from '../chatComposerTypes';
+import type { SuggestionType } from './chatComposerTypes';
 import { getCurrentWordInfo } from './getCurrentWordInfo';
 
 export const getCurrentWordAndType = (text: string, cursorPosition: number) => {
