+++ src/components/Chat/components/ChatComposer/util/suggestionRailAnimations.ts
