+++ src/components/Chat/components/EmoteSheet/EmoteSheet.tsx
@@ -1,6 +1,10 @@
+import { useCallback, useMemo, useRef, useState } from 'react';
 import { ActivityIndicator, ScrollView, View } from 'react-native';
+import { useTranslation } from 'react-i18next';
+import type { LayoutChangeEvent } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
+import type { LegendListRef } from '@legendapp/list/react-native';
 import { LegendList } from '@legendapp/list/react-native';
 
 import {
@@ -12,26 +16,18 @@ import { theme } from '@app/styles/themes';
 
 import { EmoteSearchFilter } from './EmoteSearchFilter';
 import { EmoteSheetIosBlur } from './EmoteSheetIosBlur';
-import { renderSetRailItem } from './EmoteSheetSetRailItem';
-import { emoteSheetStyles as styles } from './emoteSheetStyles';
-import type { EmotePickerItem } from './emoteSheetTypes';
-import { ProviderChip } from './ProviderChip';
-import { emoteSheetScrollActivity } from './util/emoteSheetScrollActivity';
-
-export type { EmotePickerItem };
-import { useCallback, useMemo, useRef, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import type { LayoutChangeEvent } from 'react-native';
-
-import type { LegendListRef } from '@legendapp/list/react-native';
-
 import {
   EMOTE_CELL_GAP,
   EMOTE_SHEET_DETENT,
   EMOTE_SHEET_HEADER_HEIGHT,
 } from './emoteSheetLayout';
+import { renderSetRailItem } from './EmoteSheetSetRailItem';
+import { emoteSheetStyles as styles } from './emoteSheetStyles';
+import type { EmotePickerItem } from './emoteSheetTypes';
+import { ProviderChip } from './ProviderChip';
 import { useEmoteSheet } from './useEmoteSheet';
 import type { EmoteMenuListItem } from './util/emoteMenuData';
+import { emoteSheetScrollActivity } from './util/emoteSheetScrollActivity';
 
 const keyExtractor = (item: EmoteMenuListItem) => item.key;
 const getItemType = (item: EmoteMenuListItem) => item.type;
