+++ src/components/Chat/components/EmoteSheet/util/emoteMenuData.ts
@@ -2,7 +2,7 @@ import { BrandIconName } from '@app/components/BrandIcon/brandIconRegistry';
 import type { SubscriberChannelProfile } from '@app/store/chat/types/constants';
 import type { SanitisedEmote } from '@app/types/emote';
 
-import type { EmotePickerItem } from '../EmoteSheet';
+import type { EmotePickerItem } from '../emoteSheetTypes';
 
 export type EmoteMenuProviderId = '7TV' | 'Twitch' | 'FFZ' | 'BTTV' | 'Emoji';
 export type EmoteMenuIcon =
