+++ src/components/Chat/util/__tests__/__fixtures__/resolveMessageEmoteParts.perf.fixture.ts
