+++ src/components/Chat/util/chatInputSectionTypes.ts
@@ -3,7 +3,7 @@ import type { RefObject } from 'react';
 import type { ChatConnectionFlags } from '@app/components/Chat/types/chatUiFlags';
 import type { ParsedPart } from '@app/utils/chat/parsedPart';
 
-import type { ChatComposerHandle } from './ChatComposer/ChatComposer';
+import type { ChatComposerHandle } from '../components/ChatComposer/ChatComposer';
 
 export interface ReplyToData {
   messageId: string;
