+++ src/components/Chat/util/composerSizing.ts
