+++ src/components/Chat/util/formatModerationSystemMessage/formatModerationSystemMessage.ts
