--- src/components/Chat/util/richChatMessageHelpers/normaliseUsername.ts
@@ -1,3 +1,0 @@
-export function normaliseUsername(value?: string): string {
-  return value?.trim().replace(/^@/, '').toLowerCase() ?? '';
-}
