+++ src/dev/chatHotspotBench/scenarios.ts
@@ -11,7 +11,7 @@ import { chatLineMetrics } from '@app/components/Chat/components/ChatMessage/Ric
 import {
   denseEmoteData,
   reprocessChatLines,
-} from '@app/components/Chat/util/__fixtures__/resolveMessageEmoteParts.perf.fixture';
+} from '@app/components/Chat/util/__tests__/__fixtures__/resolveMessageEmoteParts.perf.fixture';
 import { resolveMessageEmoteParts } from '@app/components/Chat/util/resolveMessageEmoteParts';
 import {
   ingestBurstMessages,
